test: prepare for Content to be broken up again
The existing Content will become a base class, from which Page
and Post classes will inherit.

I envisage Content to check the slug for validity and normalise
if not, or create a slug from the title if a slug isn't provided,
as well as implementing title and slug-based comparison, but
everything else is likely to go on the subclasses.

I didn't properly document why I combined the Post and Page
classes before, but from looking at the logs I can only guess
it was because I was struggling with the file parser.

A lot of the complexity that may have built up in the parsing
function can be put into the classes themselves, so you're not
tripping over yourself checking what kind of type you're working
with.