move url and output_path basic prop tests to BlogObject
These are the tests that check for raising from the base class and
for overriding the property by directly setting — there are still
tests on the subclasses that check that url and output_path return
the expected computed value. These will be kept to test proper
subclassing of BlogObject (or reimplementation if that's needed).