majestic: use UTC to check if Post date is in the future
As I had it before, it would just localize the machine time provided
by datetime.now but without making any adustments to the time itself.
This would cause problems if majestic is run on a machine in a more
westerly timezone than the one in which the posts were created.

(So, generally, if your timezone in settings is Europe/London and you
give a post a time of 5pm, but use majestic on a server in the eastern
US within five hours, the post will be falsely considered to be
in the future and as such a draft.)

Providing a UTC timezone to datetime.now does correctly adjust the
machine time to the current UTC time