test: don't required resolved paths
Force parse_copy_paths to resolve the source path means that the function
is dependent on being run with the correct working directory.

This should not enter into consideration - the caller should make sure that
any copy functions invoked using the paths returned are called from the
correct place (or explicitly resolved beforehand).